% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{billboard}
\alias{billboard}
\title{Song rankings for Billboard top 100 in the year 2000}
\format{
A dataset with variables:
\describe{
\item{artist}{Artist name}
\item{track}{Song name}
\item{date.enter}{Date the song entered the top 100}
\item{wk1 -- wk76}{Rank of the song in each week after it entered}
}
}
\source{
The "Whitburn" project, \url{https://waxy.org/2008/05/the_whitburn_project/},
(downloaded April 2008)
}
\usage{
billboard
}
\description{
Song rankings for Billboard top 100 in the year 2000
}
\keyword{datasets}
