% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{binarize}
\alias{binarize}
\title{Extract factor levels into logical columns}
\usage{
binarize(data, ..., prefix = "")
}
\arguments{
\item{data}{A data frame.}

\item{...}{Specification of columns to binarize. Use bare variable names.
Select all variables between x and z with \code{x:z}, exclude y with
\code{-y}. For more options, see the \link[dplyr]{select} documentation.}

\item{prefix}{An optional string to prefix the new variable names}
}
\description{
Binarize takes multiple columns containing an overall set of levels
and turns them into as many logical columns as there are levels.
These columns take the names of the levels, including an optional prefix.
}
\examples{
library(dplyr)
grades <- data.frame(
                col1 = c("A", "A", "B", "C"),
                col2 = c("D", "B", "A", "A"),
                col3 = c(NA,  "C", "D", NA))

binarize(grades, starts_with("col"), prefix = "grade_")
grades \%>\% binarize(starts_with("col"), prefix = "grade_")

}
\seealso{
\code{\link{binarize_}} for a version that uses regular evaluation
  and is suitable for programming with.
}
