% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explode.R
\name{binarize_}
\alias{binarize_}
\title{Binarize (stanndard-evaluation).n}
\usage{
binarize_(data, cols, prefix = "")
}
\arguments{
\item{data}{A data frame.}

\item{cols}{Character vector giving column names to be binarized.
Select all variables between x and z with \code{x:z}, exclude y with
\code{-y}. For more options, see the \link[dplyr]{select} documentation.}

\item{prefix}{An optional string to prefix the new variable names}
}
\description{
This is a S3 genericn.
}
\keyword{internal}
