% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divvy.R
\name{divvy}
\alias{divvy}
\title{Split a dataframe into a list of dataframes}
\usage{
divvy(data, ...)
}
\arguments{
\item{data}{A dataframe}

\item{...}{variables to split by, grouping variables are always included as
splitting variables.}
}
\value{
A list of data frames
}
\description{
`divvy`` splits a dataframe into a list of dataframes based on the splitting variables.
}
\examples{
mtcars \%>\%
 dplyr::group_by(cyl) \%>\%
 divvy()

mtcars \%>\%
 dplyr::group_by(cyl) \%>\%
 divvy(wt)

 divvy(mtcars, cyl)
}
\seealso{
\code{\link[base:split]{base::split()}}
}
