% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.r
\name{drop_na_}
\alias{drop_na_}
\title{Standard-evaluation version of \code{drop_na}.}
\usage{
drop_na_(data, relevant_cols)
}
\arguments{
\item{data}{A data frame.}

\item{relevant_cols}{Character vector of column names. If empty or \code{NULL}, all
variables are considered while dropping rows}
}
\description{
Standard-evaluation version of \code{drop_na}.
}
\keyword{internal}

