% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_}
\alias{expand_}
\title{Expand (standard evaluation).}
\usage{
expand_(data, dots, ...)
}
\arguments{
\item{data}{A data frame}

\item{expand_cols}{Character vector of column names to be expanded.}
}
\description{
This is a S3 generic.
}
\keyword{internal}

