% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_}
\alias{extract_}
\title{Standard-evaluation version of \code{extract}.}
\usage{
extract_(data, col, into, regex = "([[:alnum:]]+)", remove = TRUE,
  convert = FALSE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Name of column to split, as string.}

\item{into}{Names of new variables to create as character vector.}

\item{regex}{a regular expression used to extract the desired values.}

\item{remove}{If \code{TRUE}, remove input column from output data frame.}

\item{convert}{If \code{TRUE}, will run \code{\link{type.convert}} with
\code{as.is = TRUE} on new columns. This is useful if the component
columns are integer, numeric or logical.}

\item{...}{Other arguments passed on to \code{\link{regexec}} to control
how the regular expression is processed.}
}
\description{
This is a S3 generic.
}
\keyword{internal}
