% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{Fill in missing values.}
\usage{
fill(data, col)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Bare column name.}
}
\description{
Fills missing values in using the previous entry. This is useful in the
common output format where values are not repeated, they're recorded
each time they change.
}
\examples{
df <- data.frame(Month = 1:12, Year = c(2000, rep(NA, 11)))
df \%>\% fill(Year)
}

