% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_encounters}
\alias{fish_encounters}
\title{Fish encounters}
\format{
A dataset with variables:
\describe{
\item{fish}{Fish identifier}
\item{station}{Measurement station}
\item{seen}{Was the fish seen? (1 if yes, and true for all rows)}
}
}
\source{
Dataset provided by Myfanwy Johnston; more details at
\url{https://fishsciences.github.io/post/visualizing-fish-encounter-histories/}
}
\usage{
fish_encounters
}
\description{
Information about fish swimming down a river: each station represents an
autonomous monitor that records if a tagged fish was seen at that location.
Fish travel in one direction (migrating downstream). Information about
misses is just as important as hits, but is not directly recorded in this
form of the data.
}
\keyword{datasets}
