% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{full_set}
\alias{full_set}
\title{Create the full set of values in a character vector.}
\usage{
full_set(x, vec)
}
\arguments{
\item{x}{A character vector, has to be a subset of \code{vec}}

\item{vec}{A character vector of which \code{x} is a subset}
}
\value{
\code{vec} the full set
}
\description{
Create the full set of values in a character vector.
}
\examples{
full_set(c("Jan", "Mar", "Jun"), month.abb)
full_set(c("a", "g", "b"), letters[1:15])
}

