% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gather_q}
\alias{gather_q}
\title{Gather (standard-evaluation).}
\usage{
gather_q(data, key_col, value_col, gather_cols, na.rm = FALSE,
  convert = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{key_var,value_var}{Strings giving names of key and value columns to
create.}

\item{gather_cols}{Character vector giving column names to be gathered into
pair of key-value columns.}

\item{na.rm}{If \code{TRUE}, will remove rows from output where the
value column in \code{NA}.}

\item{convert}{If \code{TRUE} will automatically run
\code{\link{type.convert}} on the key column. This is useful if the column
names are actually numeric, integer, or logical.}
}
\description{
Gather (standard-evaluation).
}
\keyword{internal}

