% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather_vars}
\alias{gather_vars}
\title{Gather variables}
\usage{
gather_vars(data, key_col, value_col, ...)
}
\description{
This function powers \code{\link{gather}()}. It is a generic function which
allows other packages to author custom methods for \code{\link{gather}()}.
}
\examples{

gather_vars(mtcars, key, value, -vs)
}
\keyword{internal}

