% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household}
\alias{household}
\title{Household data}
\format{
A data frame with 5 rows and 5 columns:
\describe{
\item{family}{Family identifier}
\item{dob_child1}{Date of birth of first child}
\item{dob_child2}{Date of birth of second child}
\item{name_child1}{Name of first child}?
\item{name_child2}{Name of second child}
}
}
\usage{
household
}
\description{
This dataset is based on an example in
\code{vignette("datatable-reshape", package = "data.table")}
}
\keyword{datasets}
