% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep-na.R
\name{keep_na}
\alias{keep_na}
\title{Keep rows containing missing values}
\usage{
keep_na(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> Columns to inspect for
missing values. If empty, all columns are used.}
}
\description{
\code{keep_na()} retains rows where any column specified by \code{...} contains a
missing value.
}
\details{
Another way to interpret \code{keep_na()} is that it only keeps the "incomplete"
rows (where rows contain missing values). Internally, this "incompleteness" is
computed through \code{\link[vctrs:missing]{vctrs::vec_detect_missing()}}.
}
\examples{
df <- tibble(x = c(1, 2, NA), y = c("a", NA, "b"))
df \%>\% keep_na()
df \%>\% keep_na(x)

vars <- "y"
df \%>\% keep_na(x, any_of(vars))
}
