% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-code.R
\name{load_data}
\alias{load_data}
\title{Load and save plant data for use with the package}
\usage{
load_data(data)
}
\arguments{
\item{data}{A data frame to be saved with \code{usethis::use_data}.}
}
\value{
Invisibly returns \code{TRUE} on success (from use_data).
}
\description{
Saves the supplied data frame as an internal \code{.rda} file, so that it can be
used in the package. It uses \code{\link[usethis:use_data]{usethis::use_data()}} to compress and overwrite
any existing data with the same name. Note: only can be used in package development.
}
\details{
Use this only if you want to bundle or update data as part of the development
workflow. Typically not called by end users unless they want to overwrite the
existing dataset.
}
\examples{
\dontrun{
my_df <- data.frame(x = 1:10, y = rnorm(10))
load_data(my_df)
}
}
