% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-spec.R, R/pivot.R
\name{pivot_spec_long}
\alias{pivot_spec_long}
\alias{pivot_spec_wide}
\alias{pivot}
\title{Pivot a data frame from wide to long or long to wide}
\usage{
pivot_spec_long(df, cols, names_to = "name", values_to = "value")

pivot_spec_wide(df, names_from = name, values_from = value,
  sep = "_")

pivot(df, spec, na.rm = FALSE, ptypes = NULL)
}
\arguments{
\item{df}{A data frame to reshape.}

\item{spec}{A data frame defining the reshaping specification.
Must contain \code{col_name} and \code{measure} columns that are character
vectors.}

\item{na.rm}{If \code{TRUE}, will convert explicit missing values to implicit
missing values. Used only when pivotting to long.}

\item{.ptype}{A named list that optionally override the types of
measured columns. Used only when pivotting to long.}
}
\description{
\code{pivot()} is provides rectangular reshaping like \code{gather()} and \code{spread()}.
It differs primarily from existing approaches in tidyr because the details
of the reshaping a described by a data frame, \code{spec}. For simple cases, you
can construct the spec with \code{pivot_spec_long()} and \code{pivot_spec_wide()}.
See details in \code{vignette("pivot")}
}
\keyword{internal}
