% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-spec.R
\name{pivot_long_spec}
\alias{pivot_long_spec}
\alias{pivot_wide_spec}
\title{Generate pivoting specifications}
\usage{
pivot_long_spec(df, cols, names_to = "name", values_to = "value")

pivot_wide_spec(df, names_from = name, values_from = value,
  names_prefix = "", names_sep = "_")
}
\arguments{
\item{df}{A data frame}

\item{cols}{Columns to pivot into longer format. This takes a tidyselect
specification.}

\item{names_to}{A string giving the name of the column to contain the data
previously stored in the column names.}

\item{values_to}{A string giving the name of the column to contain the data
previously stored in the cell values.}

\item{names_from}{Name of the variable (or variables) that will be used to
generate column names.}

\item{values_from}{Name of the variable that will be used to generate
cell values.}

\item{names_prefix}{String added to the start of every variable name. This is
particularly useful if \code{names_from} is not a character vector.}

\item{names_sep}{If \code{names_from} contains multiple variables, this will be
used to join their values together into a single string to use as
a column name.}
}
\description{
Generate pivoting specifications
}
