% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-wide.R
\name{pivot_wide}
\alias{pivot_wide}
\alias{pivot_wide_spec}
\title{Pivot data from long to wide}
\usage{
pivot_wide(df, keys = NULL, names_from = name, values_from = value,
  names_prefix = "", names_sep = "_", values_fill = NULL,
  values_collapse = NULL, spec = NULL)

pivot_wide_spec(df, names_from = name, values_from = value,
  names_prefix = "", names_sep = "_")
}
\arguments{
\item{df}{A data frame to pivot.}

\item{keys}{Keys, a set of columns that uniquely identifies each
observation. Defaults to all columns in \code{df} except for the columns
specified in \code{names_from} and \code{values_from}.}

\item{names_from, values_from}{A pair of arguments describing which column
(or columns) to get the name of the output column (\code{name_from}), and
which column to get the cell values from (\code{values_from}).}

\item{names_prefix}{String added to the start of every variable name. This is
particularly useful if \code{names_from} is a numeric vector and you want to
create syntactic variable names.}

\item{names_sep}{If \code{names_from} contains multiple variables, this will be
used to join their values together into a single string to use as
a column name.}

\item{values_fill}{Optionally, a named list specifying what each \code{value}
should be filled in with when missing.}

\item{values_collapse}{Optionally, a named list describing how to collapse
each \code{value} if the keys are not unique.}

\item{spec}{Alternatively, instead of providing \code{cols} (and \code{names_to} and
\code{values_to}) you can parse a specification data frame. This is useful
for more complex pivots because it gives you greater control on how
metadata stored in the column names turns into columns in the result.

Must be a data frame containing character \code{.name} and \code{.value} columns.}
}
\description{
\code{pivot_wide()} "widens" data, increasing the number of columns and
decreasing the number of rows. See more details in \code{vignette("pivot")},
and see \code{\link[=pivot_long]{pivot_long()}} for the inverse transformation.
}
\details{
\code{pivot_wide} is a generalisation of the function \code{spread}, it differs from
\code{spread} in two ways:

1: It's designed to be easier to read and use. Years of user feedback has
made it clear that the name of \code{spread} is not intuitive and users
(including \code{tidyr} developers) refer to the documentation almost every time
they use the function.

2: \code{pivot_wide} has been inspired by new features in other R packages, it's
therefore more powerful and flexible than \code{spread}.

However, \code{spread} will never be deprecated. For detailed examples of how to
use \code{pivot_wide}, consult the vigentte
\code{vignette("pivot", package = "tidyr")}
}
\examples{
# Use values_fill to fill in missing values when you know what they
# represent
fish_encounters \%>\%
  pivot_wide(
    names_from = station,
    values_from = seen,
    values_fill = list(seen = 0)
  )
}
