% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-wide.R
\name{pivot_wider}
\alias{pivot_wider}
\alias{pivot_wider_spec}
\title{Pivot data from long to wide}
\usage{
pivot_wider(data, id_cols = NULL, names_from = name,
  names_prefix = "", names_sep = "_", values_from = value,
  values_fill = NULL, values_fn = NULL, spec = NULL)

pivot_wider_spec(data, names_from = name, values_from = value,
  names_prefix = "", names_sep = "_")
}
\arguments{
\item{data}{A data frame to pivot.}

\item{id_cols}{A set of columns that uniquely identifies each observation.
Defaults to all columns in \code{data} except for the columns specified in
\code{names_from} and \code{values_from}. Typically used when you have additional
variables that is directly related.}

\item{names_from, values_from}{A pair of arguments describing which column
(or columns) to get the name of the output column (\code{name_from}), and
which column (or columns) to get the cell values from (\code{values_from}).

If \code{values_from} contains multiple values, the value will be added to the
front of the output column.}

\item{names_prefix}{String added to the start of every variable name. This is
particularly useful if \code{names_from} is a numeric vector and you want to
create syntactic variable names.}

\item{names_sep}{If \code{names_from} or \code{values_from} contains multiple
variables, this will be used to join their values together into a single
string to use as a column name.}

\item{values_fill}{Optionally, a named list specifying what each \code{value}
should be filled in with when missing.}

\item{values_fn}{Optionally, a named list providing a function that will be
applied to the \code{value} in each cell in the output. You will typically
use this when the combination of \code{id_cols} and \code{value} column does not
uniquely identify an observation.}

\item{spec}{Alternatively, instead of providing \code{cols} (and \code{names_to} and
\code{values_to}) you can parse a specification data frame. This is useful
for more complex pivots because it gives you greater control on how
metadata stored in the column names turns into columns in the result.

Must be a data frame containing character \code{.name} and \code{.value} columns.}
}
\description{
\code{pivot_wider()} "widens" data, increasing the number of columns and
decreasing the number of rows. See more details in \code{vignette("pivot")},
and see \code{\link[=pivot_longer]{pivot_longer()}} for the inverse transformation.
}
\section{Compared to spread}{

\code{pivot_wide} is a generalisation of the function \code{spread}, it differs from
\code{spread} in two ways:

1: It's designed to be easier to read and use. Years of user feedback has
made it clear that the name of \code{spread} is not intuitive and users
(including \code{tidyr} developers) refer to the documentation almost every time
they use the function.

2: \code{pivot_wide} has been inspired by new features in other R packages, it's
therefore more powerful and flexible than \code{spread}.

However, \code{spread} will never be deprecated. For detailed examples of how to
use \code{pivot_wide}, consult the vigentte
\code{vignette("pivot", package = "tidyr")}
}

\examples{
# Use values_fill to fill in missing values when you know what they
# represent
fish_encounters \%>\%
  pivot_wider(
    names_from = station,
    values_from = seen,
    values_fill = list(seen = 0)
  )
}
