% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tribble}
\alias{tibble}
\alias{as_tibble}
\alias{starts_with}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}, \code{\link[tibble]{tribble}}}

  \item{tidyselect}{\code{\link[tidyselect]{starts_with}}}
}}

