% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relig_income}
\alias{relig_income}
\title{Pew religion and income survey}
\format{A dataset with variables:
\describe{
\item{religion}{Name of religion}
\item{\code{<$10k}-\code{Don\'t know/refused}}{Number of respondees with
income range in column name}
}}
\source{
Downloaded from \url{https://www.pewforum.org/religious-landscape-study/}
(downloaded November 2009)
}
\usage{
relig_income
}
\description{
Pew religion and income survey
}
\keyword{datasets}
