% Generated by roxygen2 (4.0.1): do not edit by hand
\name{separate}
\alias{separate}
\title{Separate one column into multiple columns.}
\usage{
separate(data, col, into, sep = "[^[:alnum:]]+", remove = TRUE,
  convert = FALSE, ...)
}
\arguments{
\item{col}{Bare column name.}
}
\description{
Separate one column into multiple columns.
}
\examples{
df <- data.frame(x = c("a.b", "a.d", "b.c"))
separate(df, x, c("A", "B"))
}

