% Generated by roxygen2 (4.0.1): do not edit by hand
\name{separate_}
\alias{separate_}
\title{Standard-evaluation version of \code{separate}.}
\usage{
separate_(data, col, into, sep = "[^[:alnum:]]+", remove = TRUE,
  convert = FALSE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Name of column to split, as string.}

\item{sep}{Separator between columns. Defaults to any sequence
of non-alphanumeric values.}

\item{remove}{If \code{TRUE}, remove input column from data frame.}

\item{convert}{If \code{TRUE}, will run \code{\link{type.convert}} with
\code{as.is = TRUE} on new columns. This is useful if the component
columns are integer, numeric or logical.}

\item{...}{Other arguments passed on to \code{\link{strsplit}} to control
how the regular expression is processed.}
}
\description{
Standard-evaluation version of \code{separate}.
}
\keyword{internal}

