% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate2.R
\name{separate_delim_wider}
\alias{separate_delim_wider}
\alias{separate_delim_longer}
\title{Separate a character column with a delimiter}
\usage{
separate_delim_wider(
  data,
  cols,
  into,
  delim,
  extra = c("warn", "drop", "merge"),
  fill = c("warn", "right", "left"),
  names_sep = NULL
)

separate_delim_longer(data, cols, delim)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{String columns to separate into pieces. If neeeded, you can use
tidyselect to multiple columns.}

\item{into}{Names of new variables to create as character vector.
Use \code{NA} to omit the variable in the output.}

\item{delim}{Delimer between columns, a stringr regular expression.}

\item{extra}{If \code{sep} is a character vector, this controls what
happens when there are too many pieces. There are three valid options:
\itemize{
\item "warn" (the default): emit a warning and drop extra values.
\item "drop": drop any extra values without a warning.
\item "merge": only splits at most \code{length(into)} times
}}

\item{fill}{If \code{sep} is a character vector, this controls what
happens when there are not enough pieces. There are three valid options:
\itemize{
\item "warn" (the default): emit a warning and fill from the right
\item "right": fill with missing values on the right
\item "left": fill with missing values on the left
}}

\item{names_sep}{If \code{NULL}, the default, the names will be left
as is. If a string, the outer and inner names will be pasted together using
\code{names_sep} as a separator.

If the values being unnested are unnamed and \code{names_sep} is supplied, the
inner names will be automatically generated as an increasing sequence of
integers.}
}
\description{
\code{separate_delim_wider()} produces a fixed number of new columns defined by
\code{into}. \code{separate_delim_longer()} produces a variable number of rows.
}
\examples{
# If the number of components varies, it's most natural to split into rows
df <- tibble(id = 1:4, x = c("x", "x y", "x y z", NA))
df \%>\% separate_delim_longer(x, pattern = " ")

# But it's still possible to split into rows
df \%>\% separate_delim_wider(x, c("a", "b"), delim = " ")
# You can suppress the warnings by setting extra and fill
df \%>\% separate_delim_wider(x, c("a", "b"), delim = " ", extra = "drop", fill = "right")

# You can separate multiple columns at a time
df <- tibble(id = 1:3, x = c("x", "x y", "x y z"), y = c("a", "a b", "a b c"))
df \%>\% separate_delim_longer(c(x, y), " ")
}
