% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate2.R
\name{separate_longer_delim}
\alias{separate_longer_delim}
\alias{separate_longer_fixed}
\title{Split a string into rows}
\usage{
separate_longer_delim(data, cols, delim)

separate_longer_fixed(data, cols, width, keep_empty = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{String columns to separate into pieces. If neeeded, you can use
tidyselect to multiple columns.}

\item{delim}{Delimer between columns, a stringr regular expression.}

\item{width}{Number of characters to split by.}

\item{keep_empty}{By default, you'll get \code{nchar(x) / width} rows for
each observation. If \code{nchar(x)} is zero, this means the entire input
row will be dropped from the output. If you want to preserve all rows,
use \code{keep_empty = TRUE} to replace size-0 elements with a missing value.}
}
\description{
\itemize{
\item \code{separate_longer_delim()} splits by a delimiter.
\item \code{separate_longer_fixed()} splits by a fixed width.
}
}
\examples{
df <- tibble(id = 1:4, x = c("x", "x y", "x y z", NA))
df \%>\% separate_longer_delim(x, delim = " ")

# You can separate multiple columns at a time
df <- tibble(id = 1:3, x = c("x", "x y", "x y z"), y = c("a", "a b", "a b c"))
df \%>\% separate_longer_delim(c(x, y), delim = " ")

df <- tibble(id = 1:3, x = c("ab", "def", ""))
df \%>\% separate_longer_fixed(x, 1)
df \%>\% separate_longer_fixed(x, 2)
df \%>\% separate_longer_fixed(x, 2, keep_empty = TRUE)
}
