% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate2.R
\name{separate_match}
\alias{separate_match}
\title{Separate a character column by matching patterns}
\usage{
separate_match(data, cols, patterns, match_complete = TRUE, names_sep = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{String columns to separate into pieces. If neeeded, you can use
tidyselect to multiple columns.}

\item{patterns}{A named character vector where the names given names of
new columns in the output, and the values are regular expressions.
Unnamed components will match, but not be included in the output.}

\item{match_complete}{Is the pattern required to match the entire string?}

\item{names_sep}{If \code{NULL}, the default, the names will be left
as is. If a string, the outer and inner names will be pasted together using
\code{names_sep} as a separator.

If the values being unnested are unnamed and \code{names_sep} is supplied, the
inner names will be automatically generated as an increasing sequence of
integers.}
}
\description{
Separate a character column by matching patterns
}
\examples{
df <- tibble(id = 1:2, x = c("m-123", "f_4559"))
df \%>\% separate_match(x, c(gender = ".", ".", unit = "\\\\d+"))
}
