% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smiths}
\alias{smiths}
\title{Some data about the Smith family.}
\format{A data frame with 2 rows and 5 columns.}
\usage{
smiths
}
\description{
A small demo dataset describing John and Mary Smith.
}
\keyword{datasets}
