% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread_}
\alias{spread_}
\title{Standard-evaluation version of \code{spread}.}
\usage{
spread_(data, key_col, value_col, fill = NA, convert = FALSE, drop = TRUE,
  sep = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{key_col, value_col}{Strings giving names of key and value cols.}

\item{fill}{If set, missing values will be replaced with this value. Note
that there are two types of missingness in the input: explicit missing
values (i.e. \code{NA}), and implicit missings, rows that simply aren't
present. Both types of missing value will be replaced by \code{fill}.}

\item{convert}{If \code{TRUE}, \code{\link{type.convert}} with \code{asis =
TRUE} will be run on each of the new columns. This is useful if the value
column was a mix of variables that was coerced to a string. If the class of
the value column was factor or date, note that will not be true of the new
columns that are produced, which are coerced to character before type
conversion.}

\item{drop}{If \code{FALSE}, will keep factor levels that don't appear in the
data, filling in missing combinations with \code{fill}.}

\item{sep}{If \code{NULL}, the column names will be taken from the values of
\code{key} variable. If non-\code{NULL}, the column names will be given
by "<key_name><sep><key_value>".}
}
\description{
This is a S3 generic.
}
\keyword{internal}
