% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_each.R
\name{spread_each}
\alias{spread_each}
\title{Spread a key column accross multiple sub-columns.}
\usage{
spread_each(data, key, ..., fill = NA, convert = FALSE, drop = FALSE,
  sep = ".")
}
\arguments{
\item{data}{A data frame.}

\item{key}{The bare (unquoted) name of the column whose values will be used
as column headings.}

\item{...}{the bare (unquoted) names of columns that will populate 
the cells and act as sub-columns of the values of 
\code{key}.}

\item{fill}{If set, missing values will be replaced with this value. Note
that there are two types of missingness in the input: explicit missing
values (i.e. \code{NA}), and implicit missings, rows that simply aren't
present. Both types of missing value will be replaced by \code{fill}.}

\item{convert}{If \code{TRUE}, \code{\link{type.convert}} with \code{asis =
TRUE} will be run on each of the new columns. This is useful if the value
column was a mix of variables that was coerced to a string. If the class of
the value column was factor or date, note that will not be true of the new
columns that are produced, which are coerced to character before type
conversion.}

\item{drop}{If \code{FALSE}, will keep factor levels that don't appear in the
data, filling in missing combinations with \code{fill}.}

\item{sep}{Separator that goes between the key value and "..." 
column names.}
}
\description{
This works the same as \code{\link{spread}} but with multiple 
columns simultaneously.
}
\examples{
   library(dplyr)
   # generate some data
   data <- expand.grid( x = c( 'a', 'b', 'c')
                      , y = c( 'd', 'e', 'f')
                      , .rep = 1:10
                      , stringsAsFactors=FALSE
                      ) \%>\%
           mutate( v = rnorm(90)) \%>\%
           group_by(x, y) \%>\%
           summarise(N=n(), sum=sum(v))
   # a data.frame with columns x, y, N, and sum.
            
   # Spread column y over columns N and sum
   spread_each(data, y, N, sum)
   # creates a tbl_df dataset with columns:
   #   x, d.N, d.sum, e.N, e.sum, f.N, f.sum
}
\author{
Andrew Redd \email{andrew.redd@hsc.utah.edu}
}
\seealso{
\code{\link{spread_each_}} for a version that uses regular evaluation
  and is suitable for programming with.
         \code{\link{spread}} for the the single variable version.
}

