% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_each.R
\name{spread_each_}
\alias{spread_each_}
\title{Standard evaluation version of \code{spread_each}}
\usage{
spread_each_(data, key_col, ..., .dots = list(), fill = NA,
  convert = FALSE, drop = FALSE, sep = ".")
}
\arguments{
\item{data}{A data frame.}

\item{...}{strings giving names of columns that will populate
the cells and act as sub-columns of the values of 
\code{key}.}

\item{.dots}{a character vector}

\item{fill}{If set, missing values will be replaced with this value. Note
that there are two types of missingness in the input: explicit missing
values (i.e. \code{NA}), and implicit missings, rows that simply aren't
present. Both types of missing value will be replaced by \code{fill}.}

\item{convert}{If \code{TRUE}, \code{\link{type.convert}} with \code{asis =
TRUE} will be run on each of the new columns. This is useful if the value
column was a mix of variables that was coerced to a string. If the class of
the value column was factor or date, note that will not be true of the new
columns that are produced, which are coerced to character before type
conversion.}

\item{drop}{If \code{FALSE}, will keep factor levels that don't appear in the
data, filling in missing combinations with \code{fill}.}

\item{sep}{Separator that goes between the key value and "..." 
column names.}
}
\description{
Standard evaluation version of \code{spread_each}
}
\author{
Andrew Redd \email{andrew.redd@hsc.utah.edu}
}
\seealso{
\code{\link{spread_each}} for the lazy evaluation version.
         \code{\link{spread_}} for the single variable version.
}
\keyword{internal}

