% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-params.R
\name{tidyr_data_masking}
\alias{tidyr_data_masking}
\title{Argument type: data-masking}
\description{
This page describes the \verb{<data-masking>} argument modifier which
indicates that the argument uses \strong{data masking}, a sub-type of
tidy evaluation. If you've never heard of tidy evaluation before,
start by reading \code{vignette("programming")} and
\url{https://r4ds.hadley.nz/functions.html#data-frame-functions}.
}
\section{Key techniques}{
\itemize{
\item If you want the user to supply the variable (or function of variables)
in a function argument, embrace the argument, e.g. \code{filter(df, {{ var }})}.

\if{html}{\out{<div class="sourceCode R">}}\preformatted{dist_summary <- function(df, var) \{
  df \%>\%
    summarise(n = n(), min = min(\{\{ var \}\}), max = max(\{\{ var \}\}))
\}
mtcars \%>\% dist_summary(mpg)
mtcars \%>\% group_by(cyl) \%>\% dist_summary(mpg)
}\if{html}{\out{</div>}}
\item If you have the column name as a character vector, use the \code{.data}
pronoun, e.g. \code{summarise(df, mean = mean(.data[[var]]))}.

\if{html}{\out{<div class="sourceCode R">}}\preformatted{for (var in names(mtcars)) \{
  mtcars \%>\% count(.data[[var]]) \%>\% print()
\}

lapply(names(mtcars), function(var) mtcars \%>\% count(.data[[var]]))
}\if{html}{\out{</div>}}

(Note that the contents of \code{[[}, e.g. \code{var} above, is never evaluated
in the data environment so you don't need to worry about a data-variable
called \code{var} causing problems.)
\item To suppress \verb{R CMD check} \code{NOTE}s use \code{.data$var} instead of \code{var}:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# has NOTE
df \%>\% mutate(z = x + y)

# no NOTE
df \%>\% mutate(z = .data$x + .data$y)
}\if{html}{\out{</div>}}

You'll also need to import \code{.data} from rlang with (e.g.)
\verb{@importFrom rlang .data}.
}
}

\section{Dot-dot-dot (...)}{
When this modifier is applied to \code{...}, there is one other useful technique
which solves the problem of creating a new variable with a name supplied by
the user. Use the interpolation syntax from the glue package: \code{"{var}" := expression}. (Note the use of \verb{:=} instead of \code{=} to enable this syntax).

\if{html}{\out{<div class="sourceCode">}}\preformatted{var_name <- "l100km"
mtcars \%>\% mutate("\{var_name\}" := 235 / mpg)
}\if{html}{\out{</div>}}

Note that \code{...} automatically provides indirection, so you can use it as is
(i.e. without embracing) inside a function:

\if{html}{\out{<div class="sourceCode">}}\preformatted{grouped_mean <- function(df, var, ...) \{
  df \%>\%
    group_by(...) \%>\%
    summarise(mean = mean(\{\{ var \}\}))
\}
}\if{html}{\out{</div>}}
}

\keyword{internal}
