% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest}
\alias{unnest}
\title{Unnest a list column.}
\usage{
unnest(data, col)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Name of column that needs to be unnested.}
}
\description{
If you have a list-column, this makes each element of the list it's own
row.
}
\examples{
library(dplyr)
df <- data.frame(
  x = 1:3,
  y = c("a", "d,e,f", "g,h"),
  stringsAsFactors = FALSE
)
df \%>\%
  transform(y = strsplit(y, ",")) \%>\%
  unnest(y)

# Nest and unnest are inverses
df <- data.frame(x = c(1, 1, 2), y = 3:1)
df \%>\% nest(y)
df \%>\% nest(y) \%>\% unnest(y)
}
\seealso{
\code{\link{nest}} for the inverse operation.
}

