% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest2}
\alias{unnest2}
\title{Nesting and unnesting}
\usage{
unnest2(df, col, id = NULL, keep_empty = FALSE, ptype = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{col}{Column to unchop (automatically quoted).

This should be a list-column containing generalised vectors (e.g.
any mix of \code{NULL}s, atomic vector, S3 vectors, a lists, or data frames).}

\item{id}{A string specifying giving the name of a new column which will
contain the inner names of \code{col}. If unnamed, \code{col} will instead contain
numeric indices.}

\item{keep_empty}{By default, elements of \code{col} that have size zero will
be ommitted from the output. Setting \code{keep_empty = TRUE} will ensure
that they're preserved}

\item{ptype}{Optionally, supply a prototype for the output \code{col}, overriding
the default that will be guessed from the combination of individual
values.}
}
\description{
Nesting and unnesting changing both the length and width of a data frame.
They primarily work with nested-dfs, list-cols of data frames. See \code{\link[=unchop]{unchop()}}
and \code{\link[=unpack]{unpack()}} for tools tailored for list-cols of 1d vectors.
Learn more in \code{vignette("chop-pack-nest")}.
}
\examples{
df <- tibble(
  x = 1:3,
  y = list(
    NULL,
    tibble(a = 1, b = 2),
    tibble(a = 1:3, b = 3:1)
  )
)
df \%>\% unnest2(y)
df \%>\% unnest2(y, keep_empty = TRUE)
}
