% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unnest-string.R
\name{unnest_string}
\alias{unnest_string}
\title{Unnest a collapsed column.}
\usage{
unnest_string(data, ..., sep = "[^[:alnum:]]+")
}
\arguments{
\item{data}{A data frame.}

\item{...}{Specification of columns to unnest. Use bare variable names.
Select all variables between x and z with \code{x:z}, exclude y with
\code{-y}. For more options, see the \link[dplyr]{select} documentation.}

\item{sep}{Separator delimiting collapsed values.}
}
\description{
If a variable contains observations with multiple delimited values, this
separates the values and places each one in its own row.
}
\examples{
df <- data.frame(
  x = 1:3,
  y = c("a", "d,e,f", "g,h"),
  stringsAsFactors = FALSE
)
unnest_string(df, y)
}

