% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unnest-string.R
\name{unnest_string_}
\alias{unnest_string_}
\title{Standard-evaluation version of \code{unnest_string}.}
\usage{
unnest_string_(data, unnest_cols, sep = "[^[:alnum:].]+")
}
\arguments{
\item{data}{A data frame.}

\item{unnest_cols}{Name of columns that need to be unnested.}

\item{sep}{Separator delimiting collapsed values.}
}
\description{
This is a S3 generic.
}

