% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_rent_income}
\alias{us_rent_income}
\title{US rent and income data}
\format{
A dataset with variables:
\describe{
\item{GEOID}{FIP state identifier}
\item{NAME}{Name of state}
\item{variable}{Variable name: income = median yearly income,
rent = median monthly rent}
\item{estimate}{Estimated value}
\item{moe}{90\% margin of error}
}
}
\usage{
us_rent_income
}
\description{
Captured from the 2017 American Community Survey using the tidycensus
package.
}
\keyword{datasets}
