% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_bank_pop}
\alias{world_bank_pop}
\title{Population data from the World Bank}
\format{
A dataset with variables:
\describe{
\item{country}{Three letter country code}
\item{indicator}{Indicator name: \code{SP.POP.GROW} = population growth,
\code{SP.POP.TOTL} = total population, \code{SP.URB.GROW} = urban population
growth, \code{SP.URB.TOTL} = total urban population}
\item{2000-2018}{Value for each year}
}
}
\source{
Dataset from the World Bank data bank: \url{https://data.worldbank.org}
}
\usage{
world_bank_pop
}
\description{
Data about population from the World Bank.
}
\keyword{datasets}
