% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete_}
\alias{complete_}
\title{Standard-evaluation version of \code{complete}.}
\usage{
complete_(data, cols, fill = list(), ...)
}
\arguments{
\item{data}{A data frame}

\item{cols}{Columns to expand}

\item{fill}{A named list that for each variable supplies a single value to
use instead of \code{NA} for missing combinations.}
}
\description{
This is a S3 generic.
}
\keyword{internal}
