% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.r
\name{drop_na_}
\alias{drop_na_}
\title{Standard-evaluation version of \code{drop_na}.}
\usage{
drop_na_(data, vars)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Character vector of variable names. If empty, all
variables are considered while dropping rows.}
}
\description{
This is a S3 generic.
}
\keyword{internal}

