% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_numeric}
\alias{extract_numeric}
\title{Extract numeric component of variable.}
\usage{
extract_numeric(x)
}
\arguments{
\item{x}{A character vector (or a factor).}
}
\description{
Extract numeric component of variable.
}
\examples{
extract_numeric("$1,200.34")
}

