% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_numeric}
\alias{extract_numeric}
\title{Extract numeric component of variable.}
\usage{
extract_numeric(x)
}
\arguments{
\item{x}{A character vector (or a factor).}
}
\description{
DEPRECATED: please use \code{readr::parse_number()} instead.
}
\keyword{internal}
