% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_}
\alias{fill_}
\title{Standard-evaluation version of \code{fill}.}
\usage{
fill_(data, fill_cols, .direction = c("down", "up"))
}
\arguments{
\item{data}{A data frame.}

\item{fill_cols}{Character vector of column names.}

\item{.direction}{Direction in which to fill missing values. Currently
either "down" (the default) or "up".}
}
\description{
This is a S3 generic.
}
\keyword{internal}

