% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_}
\alias{fill_}
\title{Standard-evaluation version of \code{fill}.}
\usage{
fill_(data, fill_cols)
}
\arguments{
\item{data}{A data frame.}

\item{fill_cols}{Character vector of column names.}
}
\description{
This is a S3 generic.
}
\keyword{internal}

