% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nest.R
\name{nest}
\alias{nest}
\title{Nest repeated values in a list-varible.}
\usage{
nest(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Specification of columns to nest. Use bare variable names.
Select all variables between x and z with \code{x:z}, exclude y with
\code{-y}. For more options, see the \link[dplyr]{select} documentation.}
}
\description{
Nest repeated values in a list-varible.
}
\examples{
library(dplyr)
iris \%>\% nest(-Species)
chickwts \%>\% nest(weight)
}
\seealso{
\code{\link{unnest}} for the inverse operation.
}

