% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest_}
\alias{nest_}
\title{Standard-evaluation version of \code{nest}.}
\usage{
nest_(data, key_col, nest_cols = character())
}
\arguments{
\item{data}{A data frame.}

\item{key_col}{Name of the column that will contain the nested data frames.}

\item{nest_cols}{Character vector of columns to nest.}
}
\description{
This is a S3 generic.
}
\keyword{internal}

