% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pivot_long}
\alias{pivot_long}
\alias{pivot_wide}
\title{Pivot data from wide to long or long to wide}
\usage{
pivot_long(df, cols, names_to = "name", values_to = "value",
  spec = NULL, na.rm = FALSE, ptype = list())

pivot_wide(df, names_from = name, values_from = value, sep = "_",
  spec = NULL)
}
\arguments{
\item{df}{A data frame to reshape.}

\item{cols}{Columns to pivot into longer format. This takes a tidyselect
specification.}

\item{names_to}{A string giving the name of the column to contain the data
previously stored in the column names.}

\item{values_to}{A string giving the name of the column to contain the data
previously stored in the cell values.}

\item{spec}{A data frame defining the reshaping specification.
Must contain \code{.name} and \code{.value} columns that are character
vectors. See \code{\link[=pivot_wide_spec]{pivot_wide_spec()}} and \code{\link[=pivot_long_spec]{pivot_long_spec()}} for details.}

\item{na.rm}{If \code{TRUE}, will convert explicit missing values to implicit
missing values. Used only when pivotting to long.}

\item{ptype}{A named list that optionally override the types of
measured columns. Used only when pivotting to long.}

\item{names_from}{Name of the variable (or variables) that will be used to
generate column names.}

\item{values_from}{Name of the variable that will be used to generate
cell values.}

\item{sep}{If \code{names_from} contains multiple variables, this will be
used to join their values together into a single string to use as
a column name.}
}
\description{
See details in \code{vignette("pivot")}
}
