% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-spec.R, R/pivot.R
\name{pivot_wide_spec}
\alias{pivot_wide_spec}
\alias{pivot_wide}
\title{Pivot data from long to wide}
\usage{
pivot_wide_spec(df, names_from = name, values_from = value,
  names_prefix = "", names_sep = "_")

pivot_wide(df, names_from = name, values_from = value,
  names_prefix = "", names_sep = "_", values_fill = NULL,
  spec = NULL)
}
\arguments{
\item{df}{A data frame to pivot.}

\item{names_from, values_from}{A pair of arguments describing which column
(or columns) to get the name of the output column (\code{name_from}), and
which column to get the cell values from (\code{values_from}).}

\item{names_prefix}{String added to the start of every variable name. This is
particularly useful if \code{names_from} is a numeric vector and you want to
create syntactic variable names.}

\item{names_sep}{If \code{names_from} contains multiple variables, this will be
used to join their values together into a single string to use as
a column name.}

\item{values_fill}{Optionally, a named list specifying what each \code{value}
should be filled in with when missing.}

\item{spec}{Alternatively, instead of providing \code{cols} (and \code{names_to} and
\code{values_to}) you can parse a specification data frame. This is useful
for more complex pivots because it gives you greater control on how
metadata stored in the column names turns into columns in the result.

Must be a data frame containing character \code{.name} and \code{.value} columns.}
}
\description{
\code{pivot_wide()} "widens" data, increasing the number of columns and
decreasing the number of rows. See more details in \code{vignette("pivot")},
and see \code{\link[=pivot_long]{pivot_long()}} for the inverse transformation.
}
