% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{Example dataset \code{raw_data}}
\format{
A data frame with ...
}
\usage{
raw_data
}
\description{
A demonstration dataset containing ...
}
\keyword{datasets}
