% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replace missing values}
\usage{
replace_na(data, replace = list(), ...)
}
\arguments{
\item{data}{A data frame.}

\item{replace}{A named list given the value to replace \code{NA} with
for each column.}

\item{...}{Named arguments providing values to replace \code{NA} with
for each column, or an expression based on other columns

Note that if you are replacing a column named \code{replace}
you need to use replace = list(replace = 0) argument rather than a named
argument. This is to keep \code{replace_na} backwards-compatible.}
}
\description{
Replace missing values in columns, either with a specific replacement
value or with a value (or expression) based on other columns in the
data.
}
\examples{
library(dplyr)
df <- data_frame(x = c(1, 2, NA), y = c("a", NA, "b"))

# can be given as arguments or a list
df \%>\% replace_na(x = 0, y = "unknown")
df \%>\% replace_na(list(x = 0, y = "unknown"))

# can replace columns with other columns
df <- data_frame(x = c(1, 2, NA), y = c(3, NA, 4))
df \%>\% replace_na(x = y)
df \%>\% replace_na(y = x)

df \%>\% replace_na(x = y * 10)

}
