% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate-rows.R
\name{separate_rows_}
\alias{separate_rows_}
\title{Standard-evaluation version of \code{separate_rows}.}
\usage{
separate_rows_(data, cols, sep = "[^[:alnum:].]+")
}
\arguments{
\item{data}{A data frame.}

\item{cols}{Name of columns that need to be unnested.}

\item{sep}{Separator delimiting collapsed values.}
}
\description{
This is a S3 generic.
}

