% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expand.R
\name{seq_range}
\alias{seq_range}
\title{Create an evenly spaced sequence of values from lowest to highest.}
\usage{
seq_range(x, n)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{Number of values}
}
\description{
Create an evenly spaced sequence of values from lowest to highest.
}
\examples{
seq_range(1:100, 5)
}

