% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seq_range}
\alias{seq_range}
\title{Create an evenly spaced sequence of values from highest to lowest.}
\usage{
seq_range(x, n)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{Number of values}
}
\description{
Create an evenly spaced sequence of values from highest to lowest.
}
\examples{
seq_range(1:100, 5)
}

