% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spread_}
\alias{spread_}
\title{Standard-evaluation version of \code{spread}.}
\usage{
spread_(data, key_col, value_col, fill = NA, convert = FALSE, drop = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{key_col,value_col}{Strings giving names of key and value cols.}

\item{fill}{If there isn't a value for every combination of the other
variables and the key column, this value will be substituted.}

\item{convert}{If \code{TRUE}, \code{\link{type.convert}} with
\code{asis = TRUE} will be run on each of the new columns. This is
useful if the value column was a mix of variables that was coerced to
a string.}

\item{drop}{If \code{FALSE}, will keep factor levels that don't appear
in the data, filling in missing combinations with \code{fill}.}
}
\description{
This is a S3 generic.
}

