% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{TidyR-package}
\alias{TidyR}
\alias{TidyR-package}
\title{TidyR: Tidying and Plotting Data}
\description{
This package loads in a data file and tidies it by removing any NA values (spaces) and replacing them with dashes to make it easier to work with. The package also includes a function that returns a plot of an inputted x and y.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erinweadick/TidyR}
  \item Report bugs at \url{https://github.com/erinweadick/TidyR/issues}
}

}
\author{
\strong{Maintainer}: Kate Mullins \email{mullinsk38@gmail.com}

Authors:
\itemize{
  \item Erin Weadick \email{erin.weadick.2022@mumail.ie}
  \item Laurynas Jonikas \email{laurynas.jonikas.2025@mumail.ie}
  \item Daniel McKean \email{daniel.mckean.2021@mumail.ie}
}

}
\keyword{internal}
