% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr.R
\docType{package}
\name{tidyr-package}
\alias{tidyr}
\alias{tidyr-package}
\title{tidyr: Tidy Messy Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to help to create tidy data, where each column is a variable, each row is an observation, and each cell contains a single value. 'tidyr' contains tools for changing the shape (pivoting) and hierarchy (nesting and 'unnesting') of a dataset, turning deeply nested lists into rectangular data frames ('rectangling'), and extracting values out of string columns. It also includes tools for working with missing values (both implicit and explicit).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidyr.tidyverse.org}
  \item \url{https://github.com/tidyverse/tidyr}
  \item Report bugs at \url{https://github.com/tidyverse/tidyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Davis Vaughan \email{davis@posit.co}
  \item Maximilian Girlich
}

Other contributors:
\itemize{
  \item Kevin Ushey \email{kevin@posit.co} [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
