% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr.R
\docType{package}
\name{tidyr-package}
\alias{tidyr}
\alias{tidyr-package}
\title{tidyr: Easily Tidy Data with 'spread()' and 'gather()' Functions}
\description{
An evolution of 'reshape2'. It's designed specifically for data
tidying (not general reshaping or aggregating) and works well with
'dplyr' data pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://tidyr.tidyverse.org}
  \item \url{https://github.com/tidyverse/tidyr}
  \item Report bugs at \url{https://github.com/tidyverse/tidyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
