% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unite_}
\alias{unite_}
\title{Standard-evaluation version of \code{unite}}
\usage{
unite_(data, col, from, sep = "_", remove = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Name of new column as string.}

\item{from}{Names of existing columns as character vector}

\item{sep}{Separator to use between values.}

\item{remove}{If \code{TRUE}, remove \code{col} from the data.}
}
\description{
Standard-evaluation version of \code{unite}
}
\keyword{internal}

