% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{unite_}
\alias{unite_}
\title{Standard-evaluation version of \code{unite}}
\usage{
unite_(data, col, from, sep = "_", remove = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{col}{The name of the new column, as a string or symbol.

  This argument is passed by expression and supports
  [quasiquotation][rlang::quasiquotation] (you can unquote strings
  and symbols). The name is captured from the expression with
  [rlang::quo_name()] (note that this kind of interface where
  symbols do not represent actual objects is now discouraged in the
  tidyverse; we support it here for backward compatibility).}

\item{from}{Names of existing columns as character vector}

\item{sep}{Separator to use between values.}

\item{remove}{If \code{TRUE}, remove input columns from output data frame.}
}
\description{
This is a S3 generic.
}
\keyword{internal}
