% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.R
\name{unite_}
\alias{unite_}
\title{Standard-evaluation version of \code{unite}}
\usage{
unite_(data, col, from, sep = "_", remove = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Name of new column as string.}

\item{from}{Names of existing columns as character vector}

\item{sep}{Separator to use between values.}

\item{remove}{If \code{TRUE}, remove input columns from output data frame.}
}
\description{
This is a S3 generic.
}
\keyword{internal}
