% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest_}
\alias{unnest_}
\title{Standard-evaluation version of \code{unnest}.}
\usage{
unnest_(data, unnest_cols, .drop = NA, .id = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{unnest_cols}{Name of columns that needs to be unnested.}

\item{.drop}{Should additional list columns be dropped? By default,
\code{unnest} will drop them if unnesting the specified columns requires
the rows to be duplicated.}

\item{.id}{Data frame idenfier - if supplied, will create a new column
with name \code{.id}, giving a unique identifer. This is most useful if
the list column is named.}
}
\description{
This is a S3 generic.
}
\keyword{internal}

