% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest_}
\alias{unnest_}
\title{Standard-evaluation version of \code{unnest}.}
\usage{
unnest_(data, unnest_cols)
}
\arguments{
\item{data}{A data frame.}

\item{unnest_cols}{Name of columns that needs to be unnested.}
}
\description{
This is a S3 generic.
}

