% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unnest_}
\alias{unnest_}
\title{Standard-evaluation version of \code{unnest}.}
\usage{
unnest_(data, col)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Name of column that needs to be unnested.}
}
\description{
This is a S3 generic.
}

