% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unnest-string.R
\name{unnest_string_}
\alias{unnest_string_}
\title{Standard-evaluation version of \code{unnest_string}.}
\usage{
unnest_string_(data, col, sep = "[^[:alnum:]]+")
}
\arguments{
\item{data}{A data frame.}

\item{col}{Name of column that needs to be unnested.}

\item{sep}{Separator delimiting collapsed values.}
}
\description{
This is a S3 generic.
}

